requirements.txt
# Core dependencies
python-deriv-api>=0.1.0  # Deriv API client
pandas>=1.3.0            # Data manipulation and analysis
numpy>=1.20.0            # Numerical operations
scikit-learn>=1.0.0      # Simple ML models
reactivex>=4.0.0         # For reactive programming (Observable pattern)

# Utilities
python-dotenv>=0.19.0    # Environment variable management
aiohttp>=3.8.0           # Async HTTP client/server
joblib>=1.1.0            # Model serialization

# Logging & monitoring
schedule>=1.1.0          # Task scheduling
